Failing test + gentle refactor
