+++ org.hl7.fhir.utilities/src/main/java/org/hl7/fhir/utilities/Utilities.java
@@ -22,14 +22,12 @@ import java.nio.file.StandardCopyOption;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Calendar;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -92,7 +90,7 @@ public class Utilities {
    *
    * @param word the word that is to be pluralized.
    * @return the pluralized form of the word, or the word itself if it could not be pluralized
-   * @see #singularize(Object)
+   * @see Inflector#singularize(Object)
    */
   public static String pluralizeMe(String word) {
     Inflector inf = new Inflector();
@@ -1422,30 +1420,6 @@ public class Utilities {
       byteArrays.add(Arrays.copyOfRange(array, begin, array.length));
 
     return byteArrays;
-  }
-  
-  public static String presentDuration(long duration) {
-    duration = duration / 1000000;
-    String res = "";    // ;
-    long days       = TimeUnit.MILLISECONDS.toDays(duration);
-    long hours      = TimeUnit.MILLISECONDS.toHours(duration) -
-        TimeUnit.DAYS.toHours(TimeUnit.MILLISECONDS.toDays(duration));
-    long minutes    = TimeUnit.MILLISECONDS.toMinutes(duration) -
-        TimeUnit.HOURS.toMinutes(TimeUnit.MILLISECONDS.toHours(duration));
-    long seconds    = TimeUnit.MILLISECONDS.toSeconds(duration) -
-        TimeUnit.MINUTES.toSeconds(TimeUnit.MILLISECONDS.toMinutes(duration));
-    long millis     = TimeUnit.MILLISECONDS.toMillis(duration) -
-        TimeUnit.SECONDS.toMillis(TimeUnit.MILLISECONDS.toSeconds(duration));
-
-    if (days > 0)
-      res = String.format("%dd %02d:%02d:%02d.%04d", days, hours, minutes, seconds, millis);
-    else if (hours > 0)
-      res = String.format("%02d:%02d:%02d.%04d", hours, minutes, seconds, millis);
-    else //
-      res = String.format("%02d:%02d.%04d", minutes, seconds, millis);
-//    else
-//      res = String.format("%02d.%04d", seconds, millis);
-    return res;
   }
 
   public static void unzip(InputStream zip, Path target) throws IOException {
